safari-534.53-branch/LayoutTests/editing/spelling/spelling-attribute-at-child-expected.txt

Tests if the spellchecker behavior change after the spellcheck attribute changed by the script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS childHasSpellingMarker("<div contentEditable>Foo <span spellcheck='false' id='child'>[]</span> Baz</div>") is false
PASS childHasSpellingMarker("<div contentEditable>Foo <span id='child'>[]</span> Baz</div>") is true
PASS childHasSpellingMarker("<div contentEditable>Foo <span spellcheck='true' id='child'>[]</span> Baz</div>") is true
PASS childHasSpellingMarker("<div spellcheck='false' contentEditable>Foo <span spellcheck='false' id='child'>[]</span> Baz</div>") is false
PASS childHasSpellingMarker("<div spellcheck='false' contentEditable>Foo <span id='child'>[]</span> Baz</div>") is false
PASS childHasSpellingMarker("<div spellcheck='false' contentEditable>Foo <span spellcheck='true' id='child'>[]</span> Baz</div>") is true
PASS childHasSpellingMarker("<div spellcheck='true' contentEditable>Foo <span spellcheck='false' id='child'>[]</span> Baz</div>") is false
PASS childHasSpellingMarker("<div spellcheck='true' contentEditable>Foo <span id='child'>[]</span> Baz</div>") is true
PASS childHasSpellingMarker("<div spellcheck='true' contentEditable>Foo <span spellcheck='true' id='child'>[]</span> Baz</div>") is true
PASS successfullyParsed is true

TEST COMPLETE
